refactor: 프로필과 멤버 역할 분리
